refactor(studio): fold dev wait helpers
