+++ apps/studio/scripts/wait-for-electron-build.mjs
@@ -3,12 +3,19 @@ import http from 'node:http';
 import path from 'node:path';
 import { fileURLToPath, pathToFileURL } from 'node:url';
 import { rendererDevUrl } from './renderer-dev-config.mjs';
-import { studioRendererDepsReadyFile } from './studio-dev-deps.mjs';
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = path.dirname(__filename);
 const rootDir = path.resolve(__dirname, '..');
 
+export const studioRendererDepsReadyFile = path.join(
+  rootDir,
+  '.studio-dev',
+  'renderer-deps.ready',
+);
+export const initialBuildReadyPattern =
+  /success\s+build complete,\s+watching for changes\.\.\./i;
+export const defaultRendererDependencyFiles = [studioRendererDepsReadyFile];
 export const defaultRequiredFiles = [
   studioRendererDepsReadyFile,
   path.join(rootDir, 'dist/main/main.cjs'),
@@ -75,6 +82,44 @@ export const checkRendererReady = (url) =>
     });
   });
 
+export const waitForRequiredFiles = async ({
+  requiredFiles,
+  maxWaitMs = defaultMaxWaitMs,
+  pollIntervalMs = defaultPollIntervalMs,
+  readMtime = readMtimeMs,
+  now = () => Date.now(),
+  delay = sleep,
+} = {}) => {
+  const startedAt = now();
+
+  while (now() - startedAt < maxWaitMs) {
+    if (requiredFiles.every((file) => readMtime(file) !== null)) {
+      return true;
+    }
+
+    await delay(pollIntervalMs);
+  }
+
+  return false;
+};
+
+export const waitForRendererDeps = async ({
+  requiredFiles = defaultRendererDependencyFiles,
+  maxWaitMs = defaultMaxWaitMs,
+  pollIntervalMs = defaultPollIntervalMs,
+  readMtime = readMtimeMs,
+  now = () => Date.now(),
+  delay = sleep,
+} = {}) =>
+  waitForRequiredFiles({
+    requiredFiles,
+    maxWaitMs,
+    pollIntervalMs,
+    readMtime,
+    now,
+    delay,
+  });
+
 /**
  * Poll until the required build outputs are fresh AND the renderer dev
  * server is serving a 200, or until `maxWaitMs` elapses. Dependencies are
@@ -108,6 +153,24 @@ const isDirectInvocation =
   process.argv[1] && import.meta.url === pathToFileURL(process.argv[1]).href;
 
 if (isDirectInvocation) {
+  const mode = process.argv[2] ?? 'shell-build';
+
+  if (mode === 'renderer-deps') {
+    console.log('Waiting for Midscene Studio renderer dependencies...');
+
+    const ready = await waitForRendererDeps();
+
+    if (ready) {
+      console.log('Midscene Studio renderer dependencies are ready.');
+      process.exit(0);
+    }
+
+    console.error(
+      'Timed out waiting for the Midscene Studio renderer dependencies to finish building.',
+    );
+    process.exit(1);
+  }
+
   console.log('Waiting for Midscene Studio shell build output...');
 
   const ready = await waitForBuild();
